requirements.txt
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
bcrypt==3.1.4
bleach==1.5.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
enum34==1.1.6
html5lib==0.9999999
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.1.1
Markdown==2.6.9
numpy==1.13.3
olefile==0.44
parsel==1.2.0
pika==0.11.0
Pillow==4.3.0
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.5.1
PyMySQL==0.7.11
pyOpenSSL==17.3.0
PyPDF2==1.26.0
PYSTUDY==0.4.7
pytesseract==0.1.7
PyYAML==3.12
queuelib==1.4.2
reportlab==3.4.0
requests==2.18.4
Scrapy==1.4.0
selenium==3.7.0
service-identity==17.0.0
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
tornado==4.5.2
Twisted==17.9.0
urllib3==1.22
w3lib==1.18.0
Werkzeug==0.12.2
xlwt==1.3.0
zope.interface==4.4.3
